requirements.txt
langgraph>=0.0.35
langchain>=0.1.13
langchain-openai>=0.1.3
langchain-core>=0.1.42
openai>=1.14.0
python-dotenv>=1.0.1

# Database & Vector Drivers
oracledb>=2.2.0

# Observability
langfuse==4.0.0

# Polling & Integration
apscheduler>=3.10.4
requests>=2.31.0
pydantic>=2.6.4

# Database connectors
sqlalchemy>=2.0.29
